ref: change Integer type for Long type in id fields

+++ cit-sac-back-end/src/main/java/cr/co/ctpcit/citsacbackend/data/entities/inscription/ParentGuardianStudentEntityId.java
@@ -20,11 +20,11 @@ public class ParentGuardianStudentEntityId implements Serializable {
     private static final long serialVersionUID = -3727431787214240435L;
     @NotNull
     @Column(name = "student_id", nullable = false)
-    private Integer studentId;
+    private Long studentId;
 
     @NotNull
     @Column(name = "parentguardian_id", nullable = false)
-    private Integer parentGuardianId;
+    private Long parentGuardianId;
 
     @Override
     public boolean equals(Object o) {
